+++ src/domains/ChatList/Page/index.tsx
@@ -3,12 +3,13 @@ import { useGetChatListQuery } from '@/api/chatApi.ts';
 import NoData from '@/components/common/NoData.tsx';
 import Header from '@/components/common/Layout/Header';
 import Footer from '@/components/common/Layout/Footer';
-import ChatListWrap from '@/components/ChatList/ChatListWrap.tsx';
+import LoadingIcon from '@/components/common/LoadingIcon';
 import { HeaderItem } from '@/components/common/Layout/Header/Header.style.ts';
-import { Container, Divider } from '@/components/ChatList/chatList.style.ts';
 
 import ChatIcon from '@/assets/icons/chat/chat-icon.svg?react';
-import LoadingIcon from '@/components/common/LoadingIcon';
+
+import { Container, Divider } from './page.style';
+import ChatListWrap from '../components/ChatListWrap';
 
 const ChatListPage = () => {
   const { data, isLoading } = useGetChatListQuery(null, {
@@ -49,5 +50,3 @@ const ChatListPage = () => {
   );
 };
 export default ChatListPage;
-
-// 서버에 저장되거나 불러올 때 최신순으로 주기
