+++ src/domains/ChatList/Page/page.style.ts
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const Container = styled.ul`
+  height: 100%;
+  padding: 0 20px;
+  overflow: auto;
+`;
+
+export const Divider = styled.div`
+  width: 100%;
+  background-color: var(--color-gray-300);
+  color: var(--color-gray-100);
+  padding: 6px 8px;
+  font-size: var(--font-small);
+  font-weight: var(--weight-medium);
+  border-radius: 4px;
+  margin-top: 10px;
+`;
